start to separate folders and extract error&Joi schema
